+++ NEWS.md
@@ -1,10 +1,22 @@
-# flashlight 0.3.1
+# flashlight 0.4.0
 
 ## New functionality
 
-- Added the option `m_repetitions = 1` to `light_importance`. Set to higher value to get more stable estimates of variable importance as well as standard errors.
+### Major
+
+- Added `light_interaction`, a measure of overall interaction strength per covariable based on standard deviation across c-ICE curves.
+
+- Added the option `m_repetitions = 1` to `light_importance`. Set to higher value to get more stable estimates of variable importance as well as standard errors for small data.
+
+### Minor
 
 - Added the option `digits` to `plot_counts` with default 0. Helps to format large counts or fractional counts.
+
+- Added to option `center_at` to `light_ice` in order to control if ICE curves should be centered at position "first", "middle", or "last". Similarly, added `pd_center_at` to `light_profile`.
+
+## Bug fixes
+
+- Fixed a bug related to the color legend of the `plot`-method of `light_ice`.
 
 # flashlight 0.3.0